requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
sendgrid==6.10.0
reportlab==4.0.4
requests==2.31.0
channels-redis==4.1.0
redis==5.0.1
channels==4.0.0